Add files via upload
The following is the python code used for this study and the CoppeliaSim file. The sim.py and simConst.py files in the CoppeliaSim_RL-main(bristle) folder are necessary to connect to the CoppeliaSim environment.
First you need to download and install CoppeliaSim 4.1.0 as the training environment.
If the model needs to be trained, the necessary python libraries are.
	Numpy
	Pytorch
	Matplotlib
	tensorboardX
	rlschool
First you need to open the bristle1.ttt file and the train.py file, and run the train.py file directly to start training.
The network for the algorithm is written in the model.py file for the main program to call and for easy modification.
The bristle.py file is the main modification for this programming，The following functions are defined in the bristle.py file：
	Initialisation at the start of each episode in training.
	Use the API to connect to the CoppeliaSim environment.
	Retrieve the information needed in the environment such as coordinates
	Define the target of the current foraging
	Determine if the game is over and calculate the reward obtained
	Define the distance formula for measuring the straight line distance between the target and the robot
	Define the design of 4 actions
	Define closing the environment when training is complete
Test reward cases from training are stored in the log folder and pkl files from the optimised model are stored in the bristle_model folder.